smf: fix PD lease lifecycle edges from review
- RELEASE/DECLINE crashed on ogs_assert(pd_lease.active): build the
  Reply while the lease is intact, then release; stay silent on
  Decline without a binding
- expire the lease timer via SMF_EVT_N4_TIMER on the pfcp-sm state
  machine instead of inside the timer callback (timer deletion from
  a timer callback is forbidden)
- remove the route PDRs on session removal (leaked 2 PDRs per session)
- roll back the lease when the UPF rejects the route install/remove
- T1/T2 from preferred lifetime with 64-bit math; bound the response
  buffer before writing; check IP version in the DHCPv6 check
- config: plen as int (no uint8_t wrap), reject double configuration,
  reject ranges overlapping the session subnet
- upf: zero the framed-route parse struct (uninitialized v4 tail)
- pfcp: delete the xact on PD-lease send error paths

+++ lib/pfcp/context.h
@@ -535,7 +535,7 @@ ogs_pfcp_subnet_t *ogs_pfcp_find_subnet_by_dnn(int family, const char *dnn);
 int ogs_pfcp_subnet_delegated_prefix_set(
         ogs_pfcp_subnet_t *subnet,
         const char *range_ipstr, const char *range_numbits,
-        uint8_t plen, uint32_t valid_lifetime, uint32_t preferred_lifetime);
+        int plen, uint32_t valid_lifetime, uint32_t preferred_lifetime);
 int ogs_pfcp_delegated_prefix_alloc(
         ogs_pfcp_subnet_t *subnet, uint8_t *prefix /* OGS_IPV6_LEN */);
 void ogs_pfcp_delegated_prefix_free(
